\docType{data}
\name{cavendish}
\alias{cavendish}
\title{Data collected by the 18th century physicist Henry Cavendish made 23 experimental determinations of the earth's density (in g per cubic cm)}
\value{
  A data frame with 23 rows and 1 variable.
}
\description{
  Data collected by the 18th century physicist Henry
  Cavendish made 23 experimental determinations of the
  earth's density (in g per cubic cm)
}
\examples{
data(cavendish)
}
\keyword{datasets}

