\docType{data}
\name{coagulation}
\alias{coagulation}
\title{Data on the effect of three new diets on the blood coagulation times (in seconds) in a set of animals.}
\value{
  A data frame with 18 rows and 2 variables.
}
\description{
  Data on the effect of three new diets on the blood
  coagulation times (in seconds) in a set of animals.
}
\examples{
data(coagulation)
}
\keyword{datasets}

