\docType{data}
\name{contamination}
\alias{contamination}
\title{Data on the aerobic colony counts (acc, measured in colony forming
units per millilitre) of ten selected keyboards with samples taken
from four keys per keyboard on a particular day.  The keyboards were
selected randomly from those available in the computer laboratory.}
\value{
  A data frame with 40 rows and 2 variables.
}
\description{
  Data on the aerobic colony counts (acc, measured in
  colony forming units per millilitre) of ten selected
  keyboards with samples taken from four keys per keyboard
  on a particular day.  The keyboards were selected
  randomly from those available in the computer laboratory.
}
\examples{
data(contamination)
}
\references{
  This data set was taken from an article in
  Microbiologist, June 2007.
}
\keyword{datasets}

