\name{elicitgzerofn}
\alias{elicitgzerofn}
\title{Function used in the elicitation of normal-gamma prior distribution by using quantiles for mu and tau}
\usage{
  elicitgzerofn(g, taup1, tauquantile1, taup2,
    tauquantile2)
}
\arguments{
  \item{g}{Must be strictly positive.}

  \item{taup1}{probability}

  \item{taup2}{probability}

  \item{tauquantile1}{lower taup1-th quantile for tau}

  \item{tauquantile2}{lower taup2-th quantile for tau}
}
\value{
  value
}
\description{
  Function used in the elicitation of normal-gamma prior
  distribution by using quantiles for mu and tau
}
\keyword{character}

