\docType{data}
\name{hepatitis}
\alias{hepatitis}
\title{Data on the logarithm of ornithine carbonyltransferase (a liver enzyme) in a random sample of patients suffering from acute viral hepatitis.}
\value{
  A data frame with 30 rows and 1 variable.
}
\description{
  Data on the logarithm of ornithine carbonyltransferase (a
  liver enzyme) in a random sample of patients suffering
  from acute viral hepatitis.
}
\examples{
data(hepatitis)
}
\keyword{datasets}

