\docType{data}
\name{malcolm}
\alias{malcolm}
\title{Data collected by Malcolm on the height and shoe size of a random sample of students}
\value{
  A data frame with 52 rows and 2 variables.
}
\description{
  Data collected by Malcolm on the height and shoe size of
  a random sample of students
}
\examples{
data(malcolm)
}
\keyword{datasets}

