\name{mcmcCi}
\alias{mcmcCi}
\title{Equi-tailed confidence intervals from MCMC output}
\usage{
  mcmcCi(input, level = 0.95)
}
\arguments{
  \item{input}{matrix containing MCMC realisations}

  \item{level}{confidence level required}
}
\description{
  Equi-tailed confidence intervals from MCMC output
}

