\name{mcmcSummary}
\alias{mcmcSummary}
\title{Summarise and plot tabular MCMC output}
\usage{
  mcmcSummary(mat, rows = 4, lag.max = 100, bins = 30,
    show = TRUE, plot = TRUE)
}
\arguments{
  \item{mat}{Matrix of MCMC output, where the columns
  represent variables and the rows represent iterations}

  \item{rows}{Number of variables to plot per page on the
  graphics device}

  \item{lag.max}{Maximum lag for the ACF plots}

  \item{bins}{Approximate number of bins to use for the
  histograms}

  \item{show}{If TRUE, will display numerical summaries on
  the R console}

  \item{plot}{If TRUE, will plot graphical summaries on the
  default graphics device}
}
\description{
  This function summarises and plots tabular MCMC output
  such as that generated by the function normgibbs
}
\examples{
posterior=gibbsnorm(N=1000,initial=c(10,0.25),priorparam=c(10,1/100,3,12),n=100,xbar=15,s=4.5)
mcmcSummary(posterior,rows=2,bins=10)
}
\author{
  Taken from Darren Wilkinson's package smfsb
}

